# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "legacy-hakaton-mvp0 API",
        "description": "я мурчал пока писал эту аннотацию (тихонечко)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseBase"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TokenPair"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Логин",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseBase"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TokenPair"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Неверные учётные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Текущий пользователь",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseBase"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход (инвалидация токенов)",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "refresh_token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Нет токена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Обновление access по refresh (в заголовке Bearer или в теле refresh_token)",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseBase"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TokenPair"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный/отсутствующий токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Истёк/невалиден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/yandex/url": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получить URL для авторизации в Яндекс",
                "operationId": "a1bd55532f7176747fb3d222e7e2b0cc",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseBase"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/YandexAuthUrl"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/yandex/exchange": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Обмен кода на токены Яндекс и вход",
                "operationId": "b6e9ab98ee4b8ec2dac97559cf7c0310",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/YandexExchangeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseBase"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/TokenPair"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка обмена кода/получения профайла",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseBase"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponseBase": {
                "required": [
                    "statusCode",
                    "message"
                ],
                "properties": {
                    "statusCode": {
                        "description": "Базовый ответ",
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    }
                },
                "type": "object"
            },
            "TokenPair": {
                "required": [
                    "access_token",
                    "expires_in",
                    "refresh_token",
                    "refresh_expires_in"
                ],
                "properties": {
                    "access_token": {
                        "type": "string"
                    },
                    "expires_in": {
                        "type": "integer",
                        "example": 3600
                    },
                    "refresh_token": {
                        "type": "string"
                    },
                    "refresh_expires_in": {
                        "type": "integer",
                        "example": 1209600
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "login",
                    "password"
                ],
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "ZOV"
                    },
                    "password": {
                        "type": "string",
                        "example": "GOYDASVO"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "login",
                    "password"
                ],
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "ZOV"
                    },
                    "password": {
                        "type": "string",
                        "example": "GOYDASVO"
                    }
                },
                "type": "object"
            },
            "RefreshRequest": {
                "properties": {
                    "refresh_token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1..."
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "ZOV"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2025-10-12T11:31:47.661000Z"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2025-10-12T11:31:47.661000Z"
                    },
                    "id": {
                        "type": "string",
                        "example": "68eb91a3b5bc2188710e0e92"
                    }
                },
                "type": "object"
            },
            "YandexAuthUrl": {
                "required": [
                    "auth_url"
                ],
                "properties": {
                    "auth_url": {
                        "type": "string",
                        "example": "https://oauth.yandex.ru/authorize?response_type=code&client_id=...&redirect_uri=...&scope=login:info"
                    }
                },
                "type": "object"
            },
            "YandexExchangeRequest": {
                "required": [
                    "code"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "AQAAAAA...codefromyandex..."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}